rewiring of the side nav to get the register and login to display with cleaner code
