+++ client/src/utils/API.js
@@ -1,29 +1,32 @@
-import axios from 'axios';
+import axios from "axios";
 // const router = require("express").Router();
 
 // // const registerController = require("../../controllers/registerController");
 export default {
-   // Posts registered user
-   postRegister: function(newUser) {
-     return axios.post("/api/auth/register", newUser)
-}
- };
+  // Posts registered user
+  postRegister: function(newUser) {
+    return axios.post("/api/auth/register", newUser);
+  },
 
+  postLogin: function(loginUser) {
+    return axios.post("/api/auth/login", loginUser);
+  }
+};
 
 // module.exports = router;
 
- // // Gets the book with the given id
-   // getBook: function(id) {
-   //   return axios.get("/api/books/" + id);
-   // },
-   // // Deletes the book with the given id
-   // deleteBook: function(id) {
-   //   return axios.delete("/api/books/" + id);
-   // },
-   // // Saves a book to the database
-   // saveBook: function(bookData) {
-   //   return axios.post("/api/books", bookData);
-   // }
+// // Gets the book with the given id
+// getBook: function(id) {
+//   return axios.get("/api/books/" + id);
+// },
+// // Deletes the book with the given id
+// deleteBook: function(id) {
+//   return axios.delete("/api/books/" + id);
+// },
+// // Saves a book to the database
+// saveBook: function(bookData) {
+//   return axios.post("/api/books", bookData);
+// }
 
 // const router = require("express").Router();
 // const registerController = require("../../controllers/registerController");
